Make variations without compile features
