+++ src/contact_manager/legacy/evl.rs
@@ -1,6 +1,7 @@
-use crate::generate_basic_volume_manager;
+use crate::{generate_basic_volume_manager, generate_prio_volume_manager};
 
 // With EVL, the delay due to the queue is not taken into account
 // and the updates are automatic (we do not "scan" an actual local queue),
 // we just reduce the volume available
 generate_basic_volume_manager!(EVLManager, false, true);
+generate_prio_volume_manager!(PEVLManager, false, true, 3);
